splitting into parts
