+++ .gitignore
@@ -11,3 +11,4 @@ apps/VectorMap/tilecache/*.json
 apps/_vectormap/tilecache/*.json
 apps/vectormap/tilecache/*.json
 compositions/flowtest1/index.js
+.iotdb
