+++ examples/components/iot/smartlight.js
@@ -1,0 +1,3 @@
+define.class("$ui/view", function (require){
+
+});
